Fixing amount in PriceByItem

+++ src/main/java/is/bokun/dtos/SimpleMoney.java
@@ -5,14 +5,14 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SimpleMoney {
 
-    public Long amount;
+    public Double amount;
     public String currency;
 
-    public Long getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
